+++ src/pages/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import NoOrganization from '../components/NoOrganization';
-import userService, { Organization } from '../services/user.service';
+import organizationService, { Organization } from '../services/organization.service';
 
 function Dashboard() {
   const [organizations, setOrganizations] = useState<Array<Organization>>(
@@ -8,7 +8,7 @@ function Dashboard() {
   );
 
   useEffect(() => {
-    userService.getCurrentUserOrganizations().then(setOrganizations);
+    organizationService.getCurrentUserOrganizations().then(setOrganizations);
   }, []);
 
   return (
