+++ src/services/__mocks__/organization.service.ts
@@ -8,15 +8,6 @@ const OrganizationService = {
         username: 'testUser',
       },
     }),
-
-  getOrganizationUsers: () =>
-    Promise.resolve([
-      {
-        organization_id: 1,
-        user_id: 1,
-        role_id: null,
-      },
-    ]),
   createOrganization: (name: string) =>
     Promise.resolve({
       id: 1,
@@ -33,6 +24,17 @@ const OrganizationService = {
       role_id: 2
     }),
   deleteOrganization: () => Promise.resolve(''),
+  getCurrentUserOrganizations: () =>
+    Promise.resolve([
+      {
+        id: 1,
+        name: 'testOrg',
+        owner: {
+          id: 1,
+          username: 'testUser',
+        },
+      },
+    ]),
 };
 
 export default OrganizationService;
