+++ src/services/__mocks__/user.service.ts
@@ -4,18 +4,6 @@ const UserService = {
       id: 1,
       username: 'testUser',
     }),
-
-  getCurrentUserOrganizations: () =>
-    Promise.resolve([
-      {
-        id: 1,
-        name: 'testOrg',
-        owner: {
-          id: 1,
-          username: 'testUser',
-        },
-      },
-    ]),
 };
 
 export default UserService;
