+++ src/services/__tests__/organization.service.test.ts
@@ -3,8 +3,15 @@ import organizationService from '../organization.service';
 describe('organizationService', () => {
   it('gets organization by id', async () => {
     const org = { id: 1, name: 'testOrg', owner: { id: 1, username: 'test' } };
+    const requestHeader: HeadersInit = new Headers();
+    requestHeader.set('Content-Type', 'application/json');
+
     global.fetch = jest.fn(() =>
-      Promise.resolve({ ok: true, json: () => Promise.resolve(org) } as any)
+      Promise.resolve({
+        ok: true,
+        headers: requestHeader,
+        json: () => Promise.resolve(org)
+      } as any)
     );
 
     const searchOrg = await organizationService.getOrganizationById(1);
@@ -18,13 +25,47 @@ describe('organizationService', () => {
     );
   });
 
+  it('gets the current user organizations', async () => {
+    const org = [
+      { id: 1, name: 'testOrg', owner: { id: 1, username: 'test' } },
+    ];
+    const requestHeader: HeadersInit = new Headers();
+    requestHeader.set('Content-Type', 'application/json');
+
+    global.fetch = jest.fn(() =>
+      Promise.resolve({
+        ok: true,
+        headers: requestHeader,
+        json: () => Promise.resolve(org)
+      } as any)
+    );
+
+    const currentUserOrgs = await organizationService.getCurrentUserOrganizations();
+
+    expect(currentUserOrgs).toEqual(org);
+    expect(global.fetch).toBeCalledWith(
+      expect.any(String),
+      expect.objectContaining({
+        headers: { Authorization: expect.any(String) },
+      })
+    );
+  });
+
+  afterEach(() => {
+    (global.fetch as jest.Mock).mockRestore();
+  });
+
   it('gets organization user by id', async () => {
     const organizationUsers = [
       { organization_id: 1, user_id: 1, role_id: null },
     ];
+    const requestHeader: HeadersInit = new Headers();
+    requestHeader.set('Content-Type', 'application/json');
+
     global.fetch = jest.fn(() =>
       Promise.resolve({
         ok: true,
+        headers: requestHeader,
         json: () => Promise.resolve(organizationUsers),
       } as any)
     );
@@ -48,10 +89,13 @@ describe('organizationService', () => {
       name,
       owner: { id: 0, username: 'testUser' },
     };
+    const requestHeader: HeadersInit = new Headers();
+    requestHeader.set('Content-Type', 'application/json');
 
     global.fetch = jest.fn(() =>
       Promise.resolve({
         ok: true,
+        headers: requestHeader,
         json: () => Promise.resolve(organization),
       } as any)
     );
@@ -81,10 +125,13 @@ describe('organizationService', () => {
       name,
       owner: { id: 0, username: 'testUser' },
     };
+    const requestHeader: HeadersInit = new Headers();
+    requestHeader.set('Content-Type', 'application/json');
 
     global.fetch = jest.fn(() =>
       Promise.resolve({
         ok: true,
+        headers: requestHeader,
         json: () => Promise.resolve(organization),
       } as any)
     );
@@ -113,6 +160,7 @@ describe('organizationService', () => {
     global.fetch = jest.fn(() =>
       Promise.resolve({
         ok: true,
+        headers: { 'Content-Type': 'application/json' },
         text: () => Promise.resolve(''),
       } as any)
     );
